@@ -1,13 +1,13 @@
 name: alchemist
 description: A support package that aims to make golden testing in Flutter
   easier and more streamlined.
-version: 0.7.0
+version: 0.12.1
 homepage: https://github.com/Betterment/alchemist
 repository: https://github.com/Betterment/alchemist
 
 environment:
-  sdk: ">=3.0.1 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ">=3.8.0 <4.0.0"
+  flutter: ">=3.32.0"
 
 dependencies:
   equatable: ^2.0.3
@@ -22,8 +22,10 @@ dependencies:
 dev_dependencies:
   mocktail: ^0.3.0
   mocktail_image_network: ^0.3.1
+  path: ^1.8.3
   test: ^1.21.1
-  very_good_analysis: ^3.1.0
+  version: ^3.0.2
+  very_good_analysis: ^5.1.0
 
 flutter:
   uses-material-design: true
