@@ -3,29 +3,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 class TestDropdownButton extends StatelessWidget {
-  const TestDropdownButton({Key? key}) : super(key: key);
+  const TestDropdownButton({super.key});
 
   @override
   Widget build(BuildContext context) => Center(
-        child: DropdownButton<String>(
-          value: '0',
-          items: const [
-            DropdownMenuItem<String>(
-              value: '0',
-              child: Text('0'),
-            ),
-            DropdownMenuItem<String>(
-              value: '1',
-              child: Text('1'),
-            ),
-            DropdownMenuItem<String>(
-              value: '2',
-              child: Text('2'),
-            ),
-          ],
-          onChanged: (_) {},
-        ),
-      );
+    child: DropdownButton<String>(
+      value: '0',
+      items: const [
+        DropdownMenuItem<String>(value: '0', child: Text('0')),
+        DropdownMenuItem<String>(value: '1', child: Text('1')),
+        DropdownMenuItem<String>(value: '2', child: Text('2')),
+      ],
+      onChanged: (_) {},
+    ),
+  );
 }
 
 void main() {
@@ -33,10 +24,7 @@ void main() {
     goldenTest(
       'succeeds after tapping dropdown',
       fileName: 'dropdown_smoke_test',
-      constraints: const BoxConstraints(
-        maxWidth: 200,
-        maxHeight: 250,
-      ),
+      constraints: const BoxConstraints(maxWidth: 200, maxHeight: 250),
       pumpBeforeTest: (tester) async {
         await tester.pumpAndSettle();
         await tester.tap(find.byType(DropdownButton<String>));
