@@ -1,12 +1,9 @@
-/// A support package that aims to make golden testing in Flutter easier and
-/// more streamlined.
-library alchemist;
-
 export 'src/alchemist_config.dart';
 export 'src/blocked_text_image.dart';
 export 'src/golden_test.dart';
 export 'src/golden_test_group.dart';
 export 'src/golden_test_scenario.dart';
+export 'src/golden_test_theme.dart';
 export 'src/host_platform.dart';
 export 'src/interactions.dart';
 export 'src/pumps.dart';
