@@ -7,10 +7,8 @@ import 'package:flutter/material.dart';
 /// A function that returns the path of a golden test file for a given test's
 /// [fileName]. This function's return value should include the `.png`
 /// extension.
-typedef FilePathResolver = FutureOr<String> Function(
-  String fileName,
-  String environmentName,
-);
+typedef FilePathResolver =
+    FutureOr<String> Function(String fileName, String environmentName);
 
 /// {@template alchemist_config}
 /// A configuration object that contains settings used by Alchemist for
@@ -67,13 +65,15 @@ class AlchemistConfig extends Equatable {
   /// {@macro alchemist_config}
   const AlchemistConfig({
     bool? forceUpdateGoldenFiles,
+    GoldenTestTheme? goldenTestTheme,
     ThemeData? theme,
     PlatformGoldensConfig? platformGoldensConfig,
     CiGoldensConfig? ciGoldensConfig,
-  })  : _forceUpdateGoldenFiles = forceUpdateGoldenFiles,
-        _theme = theme,
-        _platformGoldensConfig = platformGoldensConfig,
-        _ciGoldensConfig = ciGoldensConfig;
+  }) : _forceUpdateGoldenFiles = forceUpdateGoldenFiles,
+       _theme = theme,
+       _goldenTestTheme = goldenTestTheme,
+       _platformGoldensConfig = platformGoldensConfig,
+       _ciGoldensConfig = ciGoldensConfig;
 
   /// The instance of the [AlchemistConfig] in the current zone used by the
   /// `alchemist` package.
@@ -159,12 +159,7 @@ class AlchemistConfig extends Equatable {
     required AlchemistConfig config,
     required T Function() run,
   }) {
-    return runZoned<T>(
-      run,
-      zoneValues: {
-        currentConfigKey: config,
-      },
-    );
+    return runZoned<T>(run, zoneValues: {currentConfigKey: config});
   }
 
   /// Whether to force the golden tests to update the golden file.
@@ -180,6 +175,13 @@ class AlchemistConfig extends Equatable {
   /// If no [ThemeData] is provided, the default [ThemeData.light] will be used.
   ThemeData? get theme => _theme;
   final ThemeData? _theme;
+
+  /// The [GoldenTestTheme] to use when generating golden tests.
+  ///
+  /// If no [GoldenTestTheme] is provided, the default
+  /// [GoldenTestTheme.standard] will be used.
+  GoldenTestTheme? get goldenTestTheme => _goldenTestTheme;
+  final GoldenTestTheme? _goldenTestTheme;
 
   /// The configuration for human-readable golden tests running in non-CI
   /// environments.
@@ -202,12 +204,14 @@ class AlchemistConfig extends Equatable {
   AlchemistConfig copyWith({
     bool? forceUpdateGoldenFiles,
     ThemeData? theme,
+    GoldenTestTheme? goldenTestTheme,
     PlatformGoldensConfig? platformGoldensConfig,
     CiGoldensConfig? ciGoldensConfig,
   }) {
     return AlchemistConfig(
       forceUpdateGoldenFiles: forceUpdateGoldenFiles ?? _forceUpdateGoldenFiles,
       theme: theme ?? _theme,
+      goldenTestTheme: goldenTestTheme ?? _goldenTestTheme,
       platformGoldensConfig: platformGoldensConfig ?? _platformGoldensConfig,
       ciGoldensConfig: ciGoldensConfig ?? _ciGoldensConfig,
     );
@@ -224,19 +228,22 @@ class AlchemistConfig extends Equatable {
     return copyWith(
       forceUpdateGoldenFiles: other?._forceUpdateGoldenFiles,
       theme: other?._theme,
-      platformGoldensConfig:
-          platformGoldensConfig.merge(other?._platformGoldensConfig),
+      goldenTestTheme: other?._goldenTestTheme,
+      platformGoldensConfig: platformGoldensConfig.merge(
+        other?._platformGoldensConfig,
+      ),
       ciGoldensConfig: ciGoldensConfig.merge(other?._ciGoldensConfig),
     );
   }
 
   @override
   List<Object?> get props => [
-        forceUpdateGoldenFiles,
-        theme,
-        platformGoldensConfig,
-        ciGoldensConfig,
-      ];
+    forceUpdateGoldenFiles,
+    theme,
+    goldenTestTheme,
+    platformGoldensConfig,
+    ciGoldensConfig,
+  ];
 }
 
 /// {@template goldens_config}
@@ -249,12 +256,12 @@ class AlchemistConfig extends Equatable {
 /// The [enabled] flag determines whether or not these golden tests are
 /// enabled. If set to `false`, these tests will not be generated or compared.
 /// Otherwise the tests will function as normal.
-/// {@endtemplate goldens_config_enabled}
+/// {@endtemplate}
 ///
 /// {@template goldens_config_file_path_resolver}
 /// The [filePathResolver] can be used to customize the name and of the golden
 /// file.
-/// {@endtemplate goldens_config_file_path_resolver}
+/// {@endtemplate}
 ///
 /// {@template goldens_config_theme}
 /// If a [theme] is provided, it will be assigned to the [MaterialApp] created
@@ -266,14 +273,14 @@ class AlchemistConfig extends Equatable {
 /// **Note:** when [obscureText] is true, tests are always rendered
 /// in the "Ahem" font family to ensure consistent results across platforms.
 /// In other words, the font family of the [theme] will be ignored.
-/// {@endtemplate goldens_config_theme}
+/// {@endtemplate}
 ///
 /// {@template goldens_config_render_shadows}
 /// The [renderShadows] flag determines whether or not shadows are rendered in
 /// golden tests.
 /// If set to `false`, all shadows are replaced with solid color blocks.
-/// {@endtemplate goldens_config_render_shadows}
-/// {@endtemplate goldens_config}
+/// {@endtemplate}
+/// {@endtemplate}
 abstract class GoldensConfig extends Equatable {
   /// {@macro goldens_config}
   const GoldensConfig({
@@ -399,7 +406,7 @@ abstract class GoldensConfig extends Equatable {
 /// By default, [renderShadows] is set to true so platform golden images are a
 /// more accurate representation of the tested widget.
 ///
-/// {@endtemplate platform_goldens_config}
+/// {@endtemplate}
 class PlatformGoldensConfig extends GoldensConfig {
   /// {@macro platform_goldens_config}
   const PlatformGoldensConfig({
@@ -419,6 +426,7 @@ class PlatformGoldensConfig extends GoldensConfig {
           theme: theme,
           tolerance: tolerance,
         );
+
 
   @override
   String get environmentName => HostPlatform.current().operatingSystem;
@@ -476,10 +484,7 @@ class PlatformGoldensConfig extends GoldensConfig {
   }
 
   @override
-  List<Object?> get props => [
-        ...super.props,
-        platforms,
-      ];
+  List<Object?> get props => [...super.props, platforms];
 }
 
 /// {@template ci_goldens_config}
@@ -499,7 +504,7 @@ class PlatformGoldensConfig extends GoldensConfig {
 /// {@macro goldens_config_render_shadows}
 /// By default, [renderShadows] is set to false to make CI tests more stable.
 ///
-/// {@endtemplate ci_goldens_config}
+/// {@endtemplate}
 class CiGoldensConfig extends GoldensConfig {
   /// {@macro ci_goldens_config}
   const CiGoldensConfig({
