@@ -7,13 +7,17 @@ void main() {
     GoldenTestGroup buildSmokeTestGroup() {
       return GoldenTestGroup(
         children: [
-          GoldenTestScenario(
-            name: 'scenario_text',
-            child: const Text('text'),
-          ),
+          GoldenTestScenario(name: 'scenario_text', child: const Text('text')),
           GoldenTestScenario(
             name: 'scenario_button',
-            child: ElevatedButton(
+            child: TextButton(
+              style: const ButtonStyle(
+                backgroundColor: WidgetStatePropertyAll(Color(0xFF2196F3)),
+                foregroundColor: WidgetStatePropertyAll(Color(0xFFFFFFFF)),
+                shadowColor: WidgetStatePropertyAll(Color(0xFFFF0000)),
+                surfaceTintColor: WidgetStatePropertyAll(Color(0xFF00FF00)),
+                overlayColor: WidgetStatePropertyAll(Color(0xFF0000FF)),
+              ),
               onPressed: () {},
               onLongPress: () {},
               child: const Text('button'),
@@ -32,14 +36,14 @@ void main() {
     goldenTest(
       'succeeds while pressed',
       fileName: 'interactions_smoke_test_pressed',
-      whilePerforming: press(find.byType(ElevatedButton)),
+      whilePerforming: press(find.byType(TextButton)),
       builder: buildSmokeTestGroup,
     );
 
     goldenTest(
       'succeeds while long pressed',
       fileName: 'interactions_smoke_test_long_pressed',
-      whilePerforming: longPress(find.byType(ElevatedButton)),
+      whilePerforming: longPress(find.byType(TextButton)),
       builder: buildSmokeTestGroup,
     );
   });
