@@ -4,6 +4,7 @@ import 'package:alchemist/src/blocked_text_image.dart';
 import 'package:alchemist/src/golden_test_adapter.dart';
 import 'package:alchemist/src/golden_test_group.dart';
 import 'package:alchemist/src/golden_test_scenario.dart';
+import 'package:alchemist/src/golden_test_theme.dart';
 import 'package:alchemist/src/pumps.dart';
 import 'package:alchemist/src/utilities.dart';
 import 'package:flutter/cupertino.dart';
@@ -32,9 +33,7 @@ class FakeLocalizationsDelegate
   }
 
   @override
-  bool shouldReload(
-    covariant LocalizationsDelegate<WidgetsLocalizations> old,
-  ) {
+  bool shouldReload(covariant LocalizationsDelegate<WidgetsLocalizations> old) {
     return false;
   }
 
@@ -51,9 +50,7 @@ void main() {
     final base = ThemeData.fallback();
     return base.copyWith(
       textTheme: base.textTheme.copyWith(
-        bodyMedium: base.textTheme.bodyMedium!.copyWith(
-          debugLabel: '<!$key!>',
-        ),
+        bodyMedium: base.textTheme.bodyMedium!.copyWith(debugLabel: '<!$key!>'),
       ),
     );
   }
@@ -87,7 +84,7 @@ void main() {
         try {
           goldenFileExpectationFn(1, 2);
           fail('expected matchesGoldenFile to be invoked');
-        } catch (e) {
+        } on TestFailure catch (e) {
           expect(
             e,
             isA<TestFailure>().having(
@@ -204,27 +201,25 @@ void main() {
       setUp(() {
         paintingContext = MockBlockedPaintingContext();
         when(
-          () => paintingContext.paintSingleChild(
-            any(that: isA<RenderObject>()),
-          ),
+          () =>
+              paintingContext.paintSingleChild(any(that: isA<RenderObject>())),
         ).thenReturn(null);
         paintingContextBuilder = (_, __) => paintingContext;
       });
 
       testWidgets('draws on painting context', (tester) async {
         const key = ValueKey('Root');
-        await tester.pumpWidget(
-          const SizedBox.square(key: key, dimension: 50),
-        );
+        await tester.pumpWidget(const SizedBox.square(key: key, dimension: 50));
         final finder = find.byKey(key);
 
-        final uiImage =
-            await adapter.getBlockedTextImage(finder: finder, tester: tester);
+        final uiImage = await adapter.getBlockedTextImage(
+          finder: finder,
+          tester: tester,
+        );
 
         verify(
-          () => paintingContext.paintSingleChild(
-            any(that: isA<RenderObject>()),
-          ),
+          () =>
+              paintingContext.paintSingleChild(any(that: isA<RenderObject>())),
         );
 
         expect(uiImage, isA<ui.Image>());
@@ -237,65 +232,56 @@ void main() {
 
     group('pumpGoldenTest', () {
       const adapter = FlutterGoldenTestAdapter();
-      GoldenTestGroup buildGroup({
-        Widget? child,
-      }) {
+      GoldenTestGroup buildGroup({Widget? child}) {
         return GoldenTestGroup(
           children: [
             GoldenTestScenario(
               name: 'scenario',
-              child: child ??
-                  const SizedBox(
-                    width: 100,
-                    height: 100,
-                    child: Placeholder(),
-                  ),
+              child:
+                  child ??
+                  const SizedBox(width: 100, height: 100, child: Placeholder()),
             ),
           ],
         );
       }
 
-      testWidgets(
-        'pumps the provided group',
-        (tester) async {
-          await adapter.pumpGoldenTest(
-            tester: tester,
-            textScaleFactor: 1,
-            constraints: const BoxConstraints(),
-            obscureFont: false,
-            globalConfigTheme: null,
-            variantConfigTheme: null,
-            pumpBeforeTest: onlyPumpAndSettle,
-            pumpWidget: onlyPumpWidget,
-            widget: buildGroup(),
-          );
+      testWidgets('pumps the provided group', (tester) async {
+        await adapter.pumpGoldenTest(
+          tester: tester,
+          textScaleFactor: 1,
+          constraints: const BoxConstraints(),
+          obscureFont: false,
+          globalConfigTheme: null,
+          variantConfigTheme: null,
+          goldenTestTheme: null,
+          pumpBeforeTest: onlyPumpAndSettle,
+          pumpWidget: onlyPumpWidget,
+          widget: buildGroup(),
+        );
 
-          expect(find.byType(GoldenTestGroup), findsOneWidget);
-        },
-      );
+        expect(find.byType(GoldenTestGroup), findsOneWidget);
+      });
 
-      testWidgets(
-        'resizes surface to fit the tested widget',
-        (tester) async {
-          final groupKey = FlutterGoldenTestAdapter.childKey;
+      testWidgets('resizes surface to fit the tested widget', (tester) async {
+        final groupKey = FlutterGoldenTestAdapter.childKey;
 
-          await adapter.pumpGoldenTest(
-            tester: tester,
-            textScaleFactor: 1,
-            constraints: const BoxConstraints(),
-            obscureFont: false,
-            globalConfigTheme: null,
-            variantConfigTheme: null,
-            pumpBeforeTest: onlyPumpAndSettle,
-            pumpWidget: onlyPumpWidget,
-            widget: buildGroup(),
-          );
+        await adapter.pumpGoldenTest(
+          tester: tester,
+          textScaleFactor: 1,
+          constraints: const BoxConstraints(),
+          obscureFont: false,
+          globalConfigTheme: null,
+          variantConfigTheme: null,
+          goldenTestTheme: null,
+          pumpBeforeTest: onlyPumpAndSettle,
+          pumpWidget: onlyPumpWidget,
+          widget: buildGroup(),
+        );
 
-          final targetSize = tester.getSize(find.byKey(groupKey));
+        final targetSize = tester.getSize(find.byKey(groupKey));
 
-          expect(tester.view.physicalSize, targetSize);
-        },
-      );
+        expect(tester.view.physicalSize, targetSize);
+      });
 
       testWidgets('sets text scale factor to provided value', (tester) async {
         await adapter.pumpGoldenTest(
@@ -305,6 +291,7 @@ void main() {
           obscureFont: false,
           globalConfigTheme: null,
           variantConfigTheme: null,
+          goldenTestTheme: null,
           pumpBeforeTest: onlyPumpAndSettle,
           pumpWidget: onlyPumpWidget,
           widget: buildGroup(),
@@ -313,48 +300,100 @@ void main() {
         expect(tester.platformDispatcher.textScaleFactor, 2.0);
       });
 
-      testWidgets(
-        'renders FlutterGoldenTestWrapper with provided '
-        'obscure font, themes and child arguments',
-        (tester) async {
+      group('padding', () {
+        testWidgets('default padding for `GoldenTestTheme`', (tester) async {
           await adapter.pumpGoldenTest(
             tester: tester,
-            textScaleFactor: 1,
+            textScaleFactor: 2,
             constraints: const BoxConstraints(),
-            obscureFont: true,
-            globalConfigTheme: globalConfigTheme,
-            variantConfigTheme: variantConfigTheme,
+            obscureFont: false,
+            globalConfigTheme: null,
+            variantConfigTheme: null,
+            goldenTestTheme: GoldenTestTheme.standard(),
             pumpBeforeTest: onlyPumpAndSettle,
             pumpWidget: onlyPumpWidget,
             widget: buildGroup(),
           );
 
-          expect(find.byType(FlutterGoldenTestWrapper), findsOneWidget);
-          expect(
-            tester.widget(find.byType(FlutterGoldenTestWrapper)),
-            isA<FlutterGoldenTestWrapper>()
-                .having((w) => w.obscureFont, 'obscureFont', isTrue)
-                .having(
-                  (w) => w.globalConfigTheme,
-                  'globalConfigTheme',
-                  isA<ThemeData>().having(
-                    identifyTheme,
-                    'theme key',
-                    equals(identifyTheme(globalConfigTheme)),
-                  ),
-                )
-                .having(
-                  (w) => w.variantConfigTheme,
-                  'variantConfigTheme',
-                  isA<ThemeData>().having(
-                    identifyTheme,
-                    'theme key',
-                    equals(identifyTheme(variantConfigTheme)),
-                  ),
-                ),
+          final box = find.byType(Padding);
+
+          expect(box, findsNWidgets(2));
+          expect(tester.widget<Padding>(box.at(0)).padding, EdgeInsets.zero);
+          expect(tester.widget<Padding>(box.at(1)).padding, EdgeInsets.zero);
+        });
+        testWidgets('custom padding on `GoldenTestTheme`', (tester) async {
+          await adapter.pumpGoldenTest(
+            tester: tester,
+            textScaleFactor: 2,
+            constraints: const BoxConstraints(),
+            obscureFont: false,
+            globalConfigTheme: null,
+            variantConfigTheme: null,
+            goldenTestTheme: GoldenTestTheme(
+              backgroundColor: Colors.white,
+              borderColor: Colors.black,
+              padding: const EdgeInsets.all(16),
+              nameTextStyle: const TextStyle(color: Colors.black),
+            ),
+            pumpBeforeTest: onlyPumpAndSettle,
+            pumpWidget: onlyPumpWidget,
+            widget: buildGroup(),
           );
-        },
-      );
+
+          final box = find.byType(Padding);
+
+          expect(box, findsNWidgets(2));
+          expect(
+            tester.widget<Padding>(box.at(0)).padding,
+            const EdgeInsets.all(16),
+          );
+          expect(
+            tester.widget<Padding>(box.at(1)).padding,
+            const EdgeInsets.all(16),
+          );
+        });
+      });
+
+      testWidgets('renders FlutterGoldenTestWrapper with provided '
+          'obscure font, themes and child arguments', (tester) async {
+        await adapter.pumpGoldenTest(
+          tester: tester,
+          textScaleFactor: 1,
+          constraints: const BoxConstraints(),
+          obscureFont: true,
+          globalConfigTheme: globalConfigTheme,
+          variantConfigTheme: variantConfigTheme,
+          goldenTestTheme: null,
+          pumpBeforeTest: onlyPumpAndSettle,
+          pumpWidget: onlyPumpWidget,
+          widget: buildGroup(),
+        );
+
+        expect(find.byType(FlutterGoldenTestWrapper), findsOneWidget);
+        expect(
+          tester.widget(find.byType(FlutterGoldenTestWrapper)),
+          isA<FlutterGoldenTestWrapper>()
+              .having((w) => w.obscureFont, 'obscureFont', isTrue)
+              .having(
+                (w) => w.globalConfigTheme,
+                'globalConfigTheme',
+                isA<ThemeData>().having(
+                  identifyTheme,
+                  'theme key',
+                  equals(identifyTheme(globalConfigTheme)),
+                ),
+              )
+              .having(
+                (w) => w.variantConfigTheme,
+                'variantConfigTheme',
+                isA<ThemeData>().having(
+                  identifyTheme,
+                  'theme key',
+                  equals(identifyTheme(variantConfigTheme)),
+                ),
+              ),
+        );
+      });
 
       testWidgets('calls the provided pumpBeforeTest', (tester) async {
         var pumpBeforeTestCalled = false;
@@ -365,6 +404,7 @@ void main() {
           obscureFont: false,
           globalConfigTheme: null,
           variantConfigTheme: null,
+          goldenTestTheme: null,
           pumpBeforeTest: (_) async => pumpBeforeTestCalled = true,
           pumpWidget: onlyPumpWidget,
           widget: buildGroup(),
@@ -382,6 +422,7 @@ void main() {
           obscureFont: false,
           globalConfigTheme: null,
           variantConfigTheme: null,
+          goldenTestTheme: null,
           pumpBeforeTest: onlyPumpAndSettle,
           pumpWidget: (tester, widget) async {
             await onlyPumpWidget(tester, widget);
@@ -400,46 +441,36 @@ void main() {
       const key = Key('child');
 
       await tester.pumpWidget(
-        const FlutterGoldenTestWrapper(
-          child: Text(
-            'test',
-            key: key,
-          ),
-        ),
+        const FlutterGoldenTestWrapper(child: Text('test', key: key)),
       );
 
       expect(find.byKey(key), findsOneWidget);
     });
 
-    testWidgets(
-      'renders a MediaQuery '
-      'based on the current window',
-      (tester) async {
-        await tester.pumpWidget(
-          const FlutterGoldenTestWrapper(
-            child: Text('test'),
-          ),
-        );
+    testWidgets('renders a MediaQuery '
+        'based on the current window', (tester) async {
+      await tester.pumpWidget(
+        const FlutterGoldenTestWrapper(child: Text('test')),
+      );
 
-        final windowMediaQuery = MediaQueryData.fromView(
-          tester.binding.platformDispatcher.views.first,
-        );
+      final windowMediaQuery = MediaQueryData.fromView(
+        tester.binding.platformDispatcher.views.first,
+      );
 
-        expect(find.byType(MediaQuery), findsOneWidget);
-        expect(
-          tester.widget(find.byType(MediaQuery)),
-          isA<MediaQuery>().having(
-            (m) => m.data,
-            'data',
-            isA<MediaQueryData>().having(
-              (m) => m.size,
-              'size',
-              windowMediaQuery.size,
-            ),
+      expect(find.byType(MediaQuery), findsOneWidget);
+      expect(
+        tester.widget(find.byType(MediaQuery)),
+        isA<MediaQuery>().having(
+          (m) => m.data,
+          'data',
+          isA<MediaQueryData>().having(
+            (m) => m.size,
+            'size',
+            windowMediaQuery.size,
           ),
-        );
-      },
-    );
+        ),
+      );
+    });
 
     group('localizations', () {
       testWidgets(
@@ -460,36 +491,32 @@ void main() {
         },
       );
 
-      testWidgets(
-        'does not override inherited locale or localizations',
-        (tester) async {
-          // ignore: prefer_const_constructors
-          final locale = Locale('en', 'US');
+      testWidgets('does not override inherited locale or localizations', (
+        tester,
+      ) async {
+        const locale = Locale('en', 'US');
 
-          await tester.pumpWidget(
-            Localizations(
-              locale: locale,
-              delegates: [
-                FakeLocalizationsDelegate(),
-              ],
-              child: FlutterGoldenTestWrapper(
-                variantConfigTheme: variantConfigTheme,
-                globalConfigTheme: globalConfigTheme,
-                child: const Text('test'),
-              ),
+        await tester.pumpWidget(
+          Localizations(
+            locale: locale,
+            delegates: [FakeLocalizationsDelegate()],
+            child: FlutterGoldenTestWrapper(
+              variantConfigTheme: variantConfigTheme,
+              globalConfigTheme: globalConfigTheme,
+              child: const Text('test'),
             ),
-          );
+          ),
+        );
 
-          final context = tester.element(find.text('test'));
-          expect(
-            WidgetsLocalizations.of(context),
-            isA<FakeWidgetsLocalizations>(),
-          );
-          expect(Localizations.localeOf(context), same(locale));
-          expect(MaterialLocalizations.of(context), isNotNull);
-          expect(CupertinoLocalizations.of(context), isNotNull);
-        },
-      );
+        final context = tester.element(find.text('test'));
+        expect(
+          WidgetsLocalizations.of(context),
+          isA<FakeWidgetsLocalizations>(),
+        );
+        expect(Localizations.localeOf(context), same(locale));
+        expect(MaterialLocalizations.of(context), isNotNull);
+        expect(CupertinoLocalizations.of(context), isNotNull);
+      });
     });
 
     group('provides theme to child that', () {
@@ -500,180 +527,158 @@ void main() {
         );
       }
 
-      testWidgets(
-        'is set to variant theme when provided',
-        (tester) async {
-          await tester.pumpWidget(
-            Theme(
-              data: contextProvidedTheme,
-              child: FlutterGoldenTestWrapper(
-                variantConfigTheme: variantConfigTheme,
-                globalConfigTheme: globalConfigTheme,
-                child: const Text('test'),
-              ),
+      testWidgets('is set to variant theme when provided', (tester) async {
+        await tester.pumpWidget(
+          Theme(
+            data: contextProvidedTheme,
+            child: FlutterGoldenTestWrapper(
+              variantConfigTheme: variantConfigTheme,
+              globalConfigTheme: globalConfigTheme,
+              child: const Text('test'),
             ),
-          );
+          ),
+        );
 
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              isA<ThemeData>().having(
-                identifyTheme,
-                'theme key',
-                equals(identifyTheme(variantConfigTheme)),
-              ),
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            isA<ThemeData>().having(
+              identifyTheme,
+              'theme key',
+              equals(identifyTheme(variantConfigTheme)),
             ),
-          );
-        },
-      );
+          ),
+        );
+      });
 
-      testWidgets(
-        'is set to inherited theme when provided',
-        (tester) async {
-          await tester.pumpWidget(
-            Theme(
-              data: contextProvidedTheme,
-              child: FlutterGoldenTestWrapper(
-                // Do not provide a variant theme.
-                globalConfigTheme: globalConfigTheme,
-                child: const Text('test'),
-              ),
-            ),
-          );
-
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              isA<ThemeData>().having(
-                identifyTheme,
-                'theme key',
-                equals(identifyTheme(contextProvidedTheme)),
-              ),
-            ),
-          );
-        },
-      );
-
-      testWidgets(
-        'is set to global theme when provided',
-        (tester) async {
-          await tester.pumpWidget(
-            // Do not provide an inherited theme.
-            FlutterGoldenTestWrapper(
+      testWidgets('is set to inherited theme when provided', (tester) async {
+        await tester.pumpWidget(
+          Theme(
+            data: contextProvidedTheme,
+            child: FlutterGoldenTestWrapper(
               // Do not provide a variant theme.
               globalConfigTheme: globalConfigTheme,
               child: const Text('test'),
             ),
-          );
+          ),
+        );
 
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              isA<ThemeData>().having(
-                identifyTheme,
-                'theme key',
-                equals(identifyTheme(globalConfigTheme)),
-              ),
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            isA<ThemeData>().having(
+              identifyTheme,
+              'theme key',
+              equals(identifyTheme(contextProvidedTheme)),
             ),
-          );
-        },
-      );
+          ),
+        );
+      });
 
-      testWidgets(
-        'is set to fallback when no theme is provided',
-        (tester) async {
-          await tester.pumpWidget(
-            // Do not provide an inherited theme.
-            const FlutterGoldenTestWrapper(
-              // Do not provide a variant theme.
-              // Do not provide a global theme.
+      testWidgets('is set to global theme when provided', (tester) async {
+        await tester.pumpWidget(
+          // Do not provide an inherited theme.
+          FlutterGoldenTestWrapper(
+            // Do not provide a variant theme.
+            globalConfigTheme: globalConfigTheme,
+            child: const Text('test'),
+          ),
+        );
+
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            isA<ThemeData>().having(
+              identifyTheme,
+              'theme key',
+              equals(identifyTheme(globalConfigTheme)),
+            ),
+          ),
+        );
+      });
+
+      testWidgets('is set to fallback when no theme is provided', (
+        tester,
+      ) async {
+        await tester.pumpWidget(
+          // Do not provide an inherited theme.
+          const FlutterGoldenTestWrapper(
+            // Do not provide a variant theme.
+            // Do not provide a global theme.
+            child: Text('test'),
+          ),
+        );
+
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            equals(ThemeData.fallback()),
+          ),
+        );
+      });
+
+      testWidgets('has its text obscured '
+          'when obscureFont is true', (tester) async {
+        const providedFontFamily = 'providedFontFamily';
+        const expectedFontFamily =
+            GoldenTestThemeDataExtensions.obscuredTextFontFamily;
+
+        await tester.pumpWidget(
+          Theme(
+            data: ThemeData(fontFamily: providedFontFamily),
+            child: const FlutterGoldenTestWrapper(
+              obscureFont: true,
               child: Text('test'),
             ),
-          );
+          ),
+        );
 
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              equals(ThemeData.fallback()),
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            isA<ThemeData>().having(
+              (theme) => theme.textTheme.bodyMedium!.fontFamily,
+              'textTheme.bodyMedium.fontFamily',
+              equals(expectedFontFamily),
             ),
-          );
-        },
-      );
+          ),
+        );
+      });
 
-      testWidgets(
-        'has its text obscured '
-        'when obscureFont is true',
-        (tester) async {
-          const providedFontFamily = 'providedFontFamily';
-          const expectedFontFamily =
-              GoldenTestThemeDataExtensions.obscuredTextFontFamily;
+      testWidgets('has alchemist text packages stripped', (tester) async {
+        const fontFamilyName = 'fontFamilyName';
+        const providedFontFamily = 'packages/alchemist/$fontFamilyName';
 
-          await tester.pumpWidget(
-            Theme(
-              data: ThemeData(
-                fontFamily: providedFontFamily,
-              ),
-              child: const FlutterGoldenTestWrapper(
-                obscureFont: true,
-                child: Text('test'),
-              ),
+        await tester.pumpWidget(
+          Theme(
+            data: ThemeData(fontFamily: providedFontFamily),
+            child: const FlutterGoldenTestWrapper(child: Text('test')),
+          ),
+        );
+
+        expect(
+          tester.widget(findWrapperTheme()),
+          isA<Theme>().having(
+            (w) => w.data,
+            'data',
+            isA<ThemeData>().having(
+              (theme) => theme.textTheme.bodyMedium!.fontFamily,
+              'textTheme.bodyMedium.fontFamily',
+              equals(fontFamilyName),
             ),
-          );
-
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              isA<ThemeData>().having(
-                (theme) => theme.textTheme.bodyMedium!.fontFamily,
-                'textTheme.bodyMedium.fontFamily',
-                equals(expectedFontFamily),
-              ),
-            ),
-          );
-        },
-      );
-
-      testWidgets(
-        'has its text packages stripped',
-        (tester) async {
-          const fontFamilyName = 'fontFamilyName';
-          const providedFontFamily = 'packages/test_package/$fontFamilyName';
-
-          await tester.pumpWidget(
-            Theme(
-              data: ThemeData(
-                fontFamily: providedFontFamily,
-              ),
-              child: const FlutterGoldenTestWrapper(
-                child: Text('test'),
-              ),
-            ),
-          );
-
-          expect(
-            tester.widget(findWrapperTheme()),
-            isA<Theme>().having(
-              (w) => w.data,
-              'data',
-              isA<ThemeData>().having(
-                (theme) => theme.textTheme.bodyMedium!.fontFamily,
-                'textTheme.bodyMedium.fontFamily',
-                equals(fontFamilyName),
-              ),
-            ),
-          );
-        },
-      );
+          ),
+        );
+      });
     });
   });
 }
