@@ -58,6 +58,7 @@ void main() {
           obscureFont: any(named: 'obscureFont'),
           variantConfigTheme: any(named: 'variantConfigTheme'),
           globalConfigTheme: any(named: 'globalConfigTheme'),
+          goldenTestTheme: any(named: 'goldenTestTheme'),
           pumpBeforeTest: any(named: 'pumpBeforeTest'),
           pumpWidget: any(named: 'pumpWidget'),
           widget: any(named: 'widget'),
@@ -104,6 +105,7 @@ void main() {
 
     testWidgets('throws when matcher fails', (tester) async {
       FutureOr<void> matcherInvocation() {
+        // Simulate a test failure.
         // ignore: only_throw_errors
         throw TestFailure('simulated failure');
       }
@@ -126,13 +128,12 @@ void main() {
           widget: const SizedBox(),
         );
         fail('Expected goldenTestRunner.run to throw TestFailure');
-      } catch (e) {
+      } on TestFailure catch (e) {
         expect(e, isA<TestFailure>());
       }
     });
 
-    testWidgets(
-        'renderShadows sets debugDisableShadows correctly '
+    testWidgets('renderShadows sets debugDisableShadows correctly '
         'and resets it after the test has run', (tester) async {
       late final bool debugDisableShadowsDuringTestRun;
 
@@ -170,6 +171,7 @@ void main() {
           obscureFont: any(named: 'obscureFont'),
           globalConfigTheme: any(named: 'globalConfigTheme'),
           variantConfigTheme: any(named: 'variantConfigTheme'),
+          goldenTestTheme: any(named: 'goldenTestTheme'),
         ),
       ).thenAnswer((_) async {
         tester.view.physicalSize = Size.zero;
